requirements.txt
pyautogui
mss
pillow
fastapi
uvicorn
python-multipart
requests
